# Use sets instead of arrays in UnionTypeBuilder for ~10% phan run spee…