…dup on large codebase

I used [phpspy](https://github.com/adsr/phpspy/) to profile this code.

phpspy output of a run on http://etsy.com codebase before this optimization: https://dasl-.github.io/phan_before_set.svg

phpspy output of a run on http://etsy.com codebase after this optimization: https://dasl-.github.io/phan_after_set.svg

Note: these SVGs should be viewed in chrome unfortunately because firefox has an SSL error due to my github user name having a hyphen.

The "before" phpspy run showed there was a lot of `in_array` activity in `UnionTypeBuilder`. I optimized by using a Set.

Timing data shows a 10% wall time improvement:
before:
```
./vendor/bin/phan -p -j 1 -f filelist.txt  284.97s user 4.34s system 99% cpu 4:49.33 total
```
after:
```
./vendor/bin/phan -p -j 1 -f filelist.txt  256.72s user 4.19s system 99% cpu 4:20.93 total
```