@TysonAndre 2019-11-06
> Note: these SVGs should be viewed in chrome unfortunately because firefox has an SSL error due to my github user name having a hyphen.


Aside: I'm having the same issue on desktop chrome on linux ("dasl-.github.io’s server IP address could not be found.")