@TysonAndre 2019-11-06
I downloaded this from https://raw.githubusercontent.com/dasl-/dasl-.github.io/master/phan_before_set.svg - it's definitely different.

I'll run some benchmarks on this and UnionType::of when I have time