@TysonAndre 2019-11-07
So it looks like using spl_object_id as array keys (`in_array_id`) is 4x faster than SplObjectHash (`hash`) for any size (I guess it's all the extra function calls made for SplObjectStorage, including iterator_to_array)

Choosing spl_object_id as array keys is an improvement over the existing code when there are at least 8 keys, when all of the objects are unique. (Assuming that properties of $this have similar performance to local variables, anyway)

```
<?php
// In the global namespace:

function test(int $n, Closure $cb, string $label, int $iterations = 10000) {
    $values = [];
    for ($i = 0; $i < $n; $i++) {
        $values[] = new stdClass();
    }

    $start = microtime(true);
    for ($iter = 0; $iter < $iterations; $iter++) {
        $new_values = $cb($values);
    }
    $end = microtime(true);
    printf("%-12s n=%-3d Took %.6f seconds\n", $label, $n, $end - $start);
}
function main() {
    $in_array = static function (array $values) {
        $result = [];
        foreach ($values as $v) {
            if (!\in_array($v, $result, true)) {
                $result[] = $v;
            }
        }
        return $result;
    };
    $in_array_id = static function (array $values) {
        $result = [];
        foreach ($values as $v) {
            $result[spl_object_id($v)] = $v;
        }
        return array_values($result);
    };
    $hash = static function (array $values) {
        $set = new SplObjectStorage();
        foreach ($values as $v) {
            $set->attach($v);
        }
        return iterator_to_array($set);
    };
    // function test(int $n, Closure $cb, string $label, int $iterations = 1000)
    foreach ([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 500] as $size) {
        test($size, $in_array, 'in_array');
        test($size, $in_array_id, 'in_array_id');
        test($size, $hash, 'hash');
    }
}
main();
/*
in_array     n=0   Took 0.000464 seconds
in_array_id  n=0   Took 0.000568 seconds
hash         n=0   Took 0.003857 seconds
in_array     n=1   Took 0.001153 seconds
in_array_id  n=1   Took 0.001663 seconds
hash         n=1   Took 0.006638 seconds
in_array     n=2   Took 0.001622 seconds
in_array_id  n=2   Took 0.002654 seconds
hash         n=2   Took 0.009201 seconds
in_array     n=3   Took 0.002030 seconds
in_array_id  n=3   Took 0.002679 seconds
hash         n=3   Took 0.011753 seconds
in_array     n=4   Took 0.002522 seconds
in_array_id  n=4   Took 0.003132 seconds
hash         n=4   Took 0.013948 seconds
in_array     n=5   Took 0.003040 seconds
in_array_id  n=5   Took 0.003366 seconds
hash         n=5   Took 0.016240 seconds
in_array     n=6   Took 0.003592 seconds
in_array_id  n=6   Took 0.003868 seconds
hash         n=6   Took 0.017941 seconds
in_array     n=7   Took 0.004188 seconds
in_array_id  n=7   Took 0.004257 seconds
hash         n=7   Took 0.020016 seconds
in_array     n=8   Took 0.004798 seconds
in_array_id  n=8   Took 0.004666 seconds
hash         n=8   Took 0.022039 seconds
in_array     n=9   Took 0.005910 seconds
in_array_id  n=9   Took 0.005379 seconds
hash         n=9   Took 0.024692 seconds
in_array     n=10  Took 0.006626 seconds
in_array_id  n=10  Took 0.005798 seconds
hash         n=10  Took 0.026901 seconds
in_array     n=100 Took 0.230080 seconds
in_array_id  n=100 Took 0.048922 seconds
hash         n=100 Took 0.217809 seconds
in_array     n=500 Took 4.547317 seconds
in_array_id  n=500 Took 0.245257 seconds
hash         n=500 Took 1.044326 seconds
 */
```