@TysonAndre 2019-11-07
Your use case should see a performance improvement with #3477 

If it doesn't, or if you see further areas of improvement, let me know.