@TysonAndre 2019-11-07
Aside: PHP 7.2 has faster in_array($element, $array, true) due to https://github.com/php/php-src/blob/PHP-7.2.24/NEWS#L1202-L1203 . 

Since PHP 7.1 is nearing the security end of life in December, I'm not going to focus on optimizing 7.1. This change is worse until n=100 for php 7.1, but can improve with https://github.com/runkit7/runkit_object_id installed for php 7.1

- https://www.php.net/supported-versions.php

```
With php 7.1.30 NTS
» php bench.php
in_array     n=0   Took 0.000538 seconds
in_array_id  n=0   Took 0.000798 seconds
hash         n=0   Took 0.004406 seconds
in_array     n=1   Took 0.001510 seconds
in_array_id  n=1   Took 0.003264 seconds
hash         n=1   Took 0.006817 seconds
in_array     n=2   Took 0.001678 seconds
in_array_id  n=2   Took 0.004575 seconds
hash         n=2   Took 0.010045 seconds
in_array     n=3   Took 0.002158 seconds
in_array_id  n=3   Took 0.007309 seconds
hash         n=3   Took 0.012158 seconds
in_array     n=4   Took 0.002734 seconds
in_array_id  n=4   Took 0.008053 seconds
hash         n=4   Took 0.015626 seconds
in_array     n=5   Took 0.003138 seconds
in_array_id  n=5   Took 0.009646 seconds
hash         n=5   Took 0.017520 seconds
in_array     n=6   Took 0.003781 seconds
in_array_id  n=6   Took 0.011257 seconds
hash         n=6   Took 0.020555 seconds
in_array     n=7   Took 0.004479 seconds
in_array_id  n=7   Took 0.012803 seconds
hash         n=7   Took 0.021975 seconds
in_array     n=8   Took 0.005434 seconds
in_array_id  n=8   Took 0.015721 seconds
hash         n=8   Took 0.023282 seconds
in_array     n=9   Took 0.007238 seconds
in_array_id  n=9   Took 0.017861 seconds
hash         n=9   Took 0.026928 seconds
in_array     n=10  Took 0.007493 seconds
in_array_id  n=10  Took 0.020039 seconds
hash         n=10  Took 0.029916 seconds
in_array     n=100 Took 0.212604 seconds
in_array_id  n=100 Took 0.182407 seconds
hash         n=100 Took 0.222989 seconds
in_array     n=500 Took 4.304044 seconds
in_array_id  n=500 Took 0.898012 seconds
hash         n=500 Took 1.051997 seconds
```