@dasl- 2019-11-07
Tested your new approach (https://github.com/phan/phan/pull/3477) on PHP 7.3.10:

Before:
```
./vendor/bin/phan -p -j 1 -f filelist.txt  235.65s user 4.23s system 99% cpu 4:46.50 total
```

After:
```
./vendor/bin/phan -p -j 1 -f filelist.txt  235.65s user 4.23s system 99% cpu 3:59.90 total
```
`((4*60+46)-(3*60+59))/(4*60+46) = 0.164` 16.4% improvement!! Thank you!